dependecies parsing using yaml parser
